Poseidon: rename modules using poseidon parameters values
