+++ ivc/src/lib.rs
@@ -1,9 +1,19 @@
 pub mod ivc;
-pub mod poseidon;
-pub mod poseidon_params;
-pub mod poseidon_quadri;
-pub mod tradi_poseidon;
-pub mod traditional_poseidon_params;
+/// Poseidon hash function with 55 full rounds, 0 partial rounds, sbox 7, a
+/// state of 3 elements and constraints of degree 2
+pub mod poseidon_55_0_7_3_2;
+/// Poseidon hash function with 55 full rounds, 0 partial rounds, sbox 7,
+/// a state of 3 elements and constraints of degree 7
+pub mod poseidon_55_0_7_3_7;
+/// Poseidon hash function with 8 full rounds, 56 partial rounds, sbox 5, a
+/// state of 3 elements and constraints of degree 2
+pub mod poseidon_8_56_5_3_2;
+/// Poseidon parameters for 55 full rounds, 0 partial rounds, sbox 7, a state of
+/// 3 elements
+pub mod poseidon_params_55_0_7_3;
+/// Poseidon parameters for 8 full rounds, 56 partial rounds, sbox 5 and a state
+/// of 3 elements
+pub mod poseidon_params_8_56_5_3;
 
 /*
 IVC circuit
