+++ ivc/src/poseidon_55_0_7_3_2/mod.rs
@@ -1,12 +1,14 @@
+//! Specialised circuit for Poseidon where we have maximum degree 2 constraints.
+
 pub mod columns;
 pub mod interpreter;
 
 #[cfg(test)]
 mod tests {
     use crate::{
-        poseidon::{columns::PoseidonColumn, interpreter, interpreter::PoseidonParams},
-        poseidon_params,
-        poseidon_params::PlonkSpongeConstantsIVC,
+        poseidon_55_0_7_3_2::{columns::PoseidonColumn, interpreter, interpreter::PoseidonParams},
+        poseidon_params_55_0_7_3,
+        poseidon_params_55_0_7_3::PlonkSpongeConstantsIVC,
     };
     use ark_ff::UniformRand;
     use kimchi_msm::{
@@ -27,12 +29,12 @@ mod tests {
 
     impl PoseidonParams<Fp, STATE_SIZE, NB_FULL_ROUND> for PoseidonBN254Parameters {
         fn constants(&self) -> [[Fp; STATE_SIZE]; NB_FULL_ROUND] {
-            let rc = &poseidon_params::static_params().round_constants;
+            let rc = &poseidon_params_55_0_7_3::static_params().round_constants;
             std::array::from_fn(|i| std::array::from_fn(|j| Fp::from(rc[i][j])))
         }
 
         fn mds(&self) -> [[Fp; STATE_SIZE]; STATE_SIZE] {
-            let mds = &poseidon_params::static_params().mds;
+            let mds = &poseidon_params_55_0_7_3::static_params().mds;
             std::array::from_fn(|i| std::array::from_fn(|j| Fp::from(mds[i][j])))
         }
     }
@@ -73,17 +75,17 @@ mod tests {
                 let exp_output: Vec<Fp> = {
                     let mut state: Vec<Fp> = vec![x, y, z];
                     poseidon_block_cipher::<Fp, PlonkSpongeConstantsIVC>(
-                        poseidon_params::static_params(),
+                        poseidon_params_55_0_7_3::static_params(),
                         &mut state,
                     );
                     state
                 };
                 let x_col: PoseidonColumn<STATE_SIZE, NB_FULL_ROUND> =
-                    PoseidonColumn::Round(NB_FULL_ROUND - 1, 0);
+                    PoseidonColumn::Round(NB_FULL_ROUND - 1, 4);
                 let y_col: PoseidonColumn<STATE_SIZE, NB_FULL_ROUND> =
-                    PoseidonColumn::Round(NB_FULL_ROUND - 1, 1);
+                    PoseidonColumn::Round(NB_FULL_ROUND - 1, 9);
                 let z_col: PoseidonColumn<STATE_SIZE, NB_FULL_ROUND> =
-                    PoseidonColumn::Round(NB_FULL_ROUND - 1, 2);
+                    PoseidonColumn::Round(NB_FULL_ROUND - 1, 14);
                 assert_eq!(witness_env.read_column(x_col), exp_output[0]);
                 assert_eq!(witness_env.read_column(y_col), exp_output[1]);
                 assert_eq!(witness_env.read_column(z_col), exp_output[2]);
@@ -125,15 +127,10 @@ mod tests {
             interpreter::apply_permutation(&mut constraint_env, &PoseidonBN254Parameters);
             let constraints = constraint_env.get_constraints();
 
-            // Constraints properties check. For this test, we do have 165 constraints
-            assert_eq!(constraints.len(), STATE_SIZE * NB_FULL_ROUND);
-            // Maximum degree of the constraints
-            assert_eq!(constraints.iter().map(|c| c.degree(1, 0)).max().unwrap(), 7);
-            // We only have degree 7 constraints
-            constraints
-                .iter()
-                .map(|c| c.degree(1, 0))
-                .for_each(|d| assert_eq!(d, 7));
+            // We have 825 constraints in total
+            assert_eq!(constraints.len(), 5 * STATE_SIZE * NB_FULL_ROUND);
+            // Maximum degree of the constraints is 2
+            assert_eq!(constraints.iter().map(|c| c.degree(1, 0)).max().unwrap(), 2);
 
             constraints
         };
